refactored customRenderers folder structure out of utils
