+++ src/components/Wiki/WikiPage/CustomRenderers/customTableRender.tsx
@@ -7,6 +7,11 @@ export const customTableRenderer = ({
 }: React.PropsWithChildren<
   ComponentPropsWithoutRef<'table'> & ReactMarkdownProps
 >) => {
-  const table = React.createElement(props.node.tagName, props, children)
-  return React.createElement('div', { className: 'table-container' }, table)
+  // const table = React.createElement(props.node.tagName, props, children)
+  // return React.createElement('div', { className: 'table-container' }, table)
+  return (
+    <div className="table-container">
+      <table {...props}>{children}</table>
+    </div>
+  )
 }
